pytorch-flask/requirements.txt
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
Pillow==9.2.0
requests==2.28.1
torch==1.11.0+cpu 
torchvision==0.12.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
